More conditional spectrum logic in cond_spectra
