+++ openquake/calculators/conditional_spectrum.py
@@ -25,7 +25,8 @@ import numpy
 from openquake.baselib import general, performance
 from openquake.commonlib.calc import compute_hazard_maps
 from openquake.hazardlib.imt import from_string
-from openquake.hazardlib.contexts import read_cmakers, outdict
+from openquake.hazardlib.contexts import read_cmakers
+from openquake.hazardlib.calc.cond_spectra import get_cs_contrib, outdict
 from openquake.calculators import base
 
 U16 = numpy.uint16
@@ -119,7 +120,7 @@ class ConditionalSpectrumCalculator(base.HazardCalculator):
             cmaker = self.cmakers[gid]
             ctxs = cmaker.read_ctxs(dstore, slice(start, stop))
             for ctx in ctxs:
-                out += cmaker.get_cs_contrib(ctx, imti, self.imls, oq.poes)
+                out += get_cs_contrib(cmaker, ctx, imti, self.imls, oq.poes)
 
         # Apply weights and get two dictionaries with integer keys
         # (corresponding to the rlz ID) and array values
@@ -135,8 +136,8 @@ class ConditionalSpectrumCalculator(base.HazardCalculator):
             cmaker = self.cmakers[gid]
             ctxs = cmaker.read_ctxs(dstore, slice(start, stop))
             for ctx in ctxs:
-                res = cmaker.get_cs_contrib(ctx, imti, self.imls, oq.poes,
-                                            cwmean[0])
+                res = get_cs_contrib(cmaker, ctx, imti, self.imls, oq.poes,
+                                     cwmean[0])
                 for g in res:
                     out[g][:, :, 2] += res[g][:, :, 2]  # STDDEV
         return out
